+++ packages/terra-application/src/terra-dev-site/app/Demo.5.app.jsx
@@ -3,7 +3,7 @@ import DemoAppIndex from './demo/DemoAppIndex';
 
 export default () => (
   // eslint-disable-next-line react/forbid-dom-props
-  <div style={{ height: '900px' }}>
+  <div style={{ height: '700px' }}>
     <DemoAppIndex />
   </div>
 );